moving debugStream to dedicated file. updating defaults to be more sane
