+++ src/sst/core/impl/interactive/Makefile.inc
@@ -7,5 +7,7 @@ sst_core_sources += \
 	impl/interactive/simpleDebug.cc \
 	impl/interactive/simpleDebug.h \
 	impl/interactive/cmdLineEditor.cc \
-	impl/interactive/cmdLineEditor.h
+	impl/interactive/cmdLineEditor.h \
+	impl/interactive/debugStream.h  \
+	impl/interactive/debugStream.cc
